Fix data types decimal and double

+++ src/Client/esWMS.Client/ViewModels/WarehouseEnvironment/Location/CreateLocationVM.cs
@@ -8,10 +8,10 @@
         public int Level { get; set; } = 1;
         public int Cell { get; set; } = 1;
         public decimal Capacity { get; set; } = 1;
-        public decimal? MaxLength { get; set; }
-        public decimal? MaxWidth { get; set; }
-        public decimal? MaxHeight { get; set; }
-        public decimal? MaxWeight { get; set; }
+        public double? MaxLength { get; set; }
+        public double? MaxWidth { get; set; }
+        public double? MaxHeight { get; set; }
+        public double? MaxWeight { get; set; }
         public string? DefaultMediaTypeId { get; set; }
     }
 }
