+++ src/Client/esWMS.Client/ViewModels/WarehouseEnvironment/WarehouseStockVM.cs
@@ -6,8 +6,8 @@
         public string ProductName { get; set; } = null!;
         public string CategoryId { get; set; } = null!;
         public string CategoryName { get; set; } = null!;
-        public int Quantity { get; set; }
-        public int BlockedQuantity { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal BlockedQuantity { get; set; }
         public decimal Value { get; set; }
     }
 }
