+++ src/Server/esWMS.Domain/Models/WarehouseStock.cs
@@ -8,8 +8,8 @@ namespace esWMS.Domain.Models
         public string ProductName { get; set; } = null!;
         public string CategoryId { get; set; } = null!;
         public string CategoryName { get; set; } = null!;
-        public int Quantity { get; set; }
-        public int BlockedQuantity { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal BlockedQuantity { get; set; }
         public decimal Value { get; set; }
 
         public Product? Product { get; set; }
